refactor(#15): BookSearchedItem 컴포넌트 코드 위치 변경
