+++ src/components/molecules/BookSearchedItem/style.ts
